+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [2.0.0-beta.3] - 2019-03-09
+### Added
+- Added eslint-plugin-jsx-a11y for accesibility
+- Added rule for no duplicated props
+
 ## [2.0.0-beta.2] - 2019-03-09
 ### Fixed
 - Fixed error with typescript preset